Change how the output buffer works to handle overlapped output better
Extend the output buffer to be four times the block size. This allows
enough room for the two resampled time-domain buffers overlapped with
one another. Output any samples from the output buffer that fit into the
channelData buffer, and store any remaining samples of the output buffer
into the first part of the buffer to be output during the next block.
This ensures that all output is successfully processed at some point, be
it during this block or the next.